requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
requests
sentence-transformers
faiss-cpu
numpy